componentize dashboard summary and table to make them more easily configurable for different dashboards
